Assignment for the project to predict traffic accident severity by machine learning(week-1)
A description of the data and how it will be used to solve the problem
1) Data Source:
Data provided by the Seattle Department of Transportation (SDOT) on vehicle accidents along with its severity is used to derive insights and patterns on how and when these accidents have taken place with the environmental factors like weather, road conditions etc. The dataset consists of 40 columns  having different kinds of data like, collision severity, road conditions, number of people involved, location of collision, weather etc. 
2)Data understanding
According to original data, many columns are not fit for analysis and will not use for model, and these columns will be dropped based on data understanding.
Most of  features are object types, and they need become numerical type when used for model.
3)Features selection and data handling
According to problem statement and background, the predictor or target variable will be 'SEVERITYCODE' because it is used measure the severity of an accident from 0 to 5 within the dataset. Attributes used to weigh the severity of an accident are : ‘ADDRTYPE’, JUNCTIONTYPE’, ‘WEATHER’, ‘ROADCOND’ and ‘LIGHTCOND’.
For these categorical variables, we need convert them to numerical values for logical model  of machine learning.

For the exploratory data analysis, numerical variables such as (PERSONCOUNT, VEHCOUNT, PEDCOUNT, PEDCYLCOUNT) will be chosed.

4) Methodology:
We will use the popular data package: numpy, pandas and sklearn, for data analysis, to see the correlationship between different features and Severitycode, and selected the most important features for predictive modeling
of accident severity.
Three machine learning models will be used for the accident severity prediction:
a.K Nearest Neighbour (KNN),it will help us predict the severity code of an outcome by finding the most similar to data point within k distance.
b. Decision tree,it will give us a layout of all possible outcomes so we can fully analyze the concequences of a decision.
c.Logistic Regression. For this project, our dataset will provide us two severity code outcomes, our model will only predict one of those two classes. This makes our data binary, which is perfect to use with logistic regression.